src/uarm/scripts/Uarm_teleop/Zhonglin_servo/requirements.txt
pyserial>=3.4
numpy>=1.19.0
